+++ projects/composablekernel/include/ck/tensor_operation/gpu/grid/gridwise_batched_gemm_gemm_wmma_cshuffle_v3.hpp
@@ -127,6 +127,7 @@ struct GridwiseBatchedGemmGemm_wmma_cshuffle_v3
         WaveSize1 == 32 || WaveSize1 == 64,
         "Misconfigured wave parameters: BlockSize / (MWaves * NWaves) != 32/64 threads per wave");
 
+<<<<<<< HEAD
     static constexpr index_t KPerWmmaBlk =
         WmmaSelector<ADataType, B0DataType, Acc0DataType, MPerWmma, LPerWmma>::selected_wmma
             .k_per_blk;
@@ -141,6 +142,15 @@ struct GridwiseBatchedGemmGemm_wmma_cshuffle_v3
         KInner *
         WmmaSelector<ADataType, B0DataType, Acc0DataType, MPerWmma, LPerWmma>::selected_wmma
             .k_per_wmma;
+=======
+    // TODO: Each GEMM should have own KPack
+    static constexpr index_t KPack = math::max(
+        math::lcm(math::lcm(AK1Value, L1Value), BK1Value),
+        WmmaSelector<ADataType, B0DataType, Acc0DataType, MPerWmma, LPerWmma>::selected_wmma
+            .k_per_wmma,
+        WmmaSelector<ADataType, B1DataType, Acc1DataType, MPerWmma, NPerWmma>::selected_wmma
+            .k_per_wmma);
+>>>>>>> origin/gfx1250
 
     using ThisThreadBlock = ThisThreadBlock<BlockSize>;