+++ projects/hiptensor/library/src/contraction/contraction_selection.hpp
@@ -34,26 +34,26 @@ namespace hiptensor
     class ContractionSolution;
     struct PerfMetrics;
 
-    hiptensorStatus_t bruteForceModel(ContractionSolution**                    winner,
-                                      std::vector<ContractionSolution*>&       candidates,
-                                      hiptensorDataType_t                      typeA,
-                                      std::vector<std::size_t> const&          a_ms_ks_lengths,
-                                      std::vector<std::size_t> const&          a_ms_ks_strides,
-                                      std::vector<int32_t> const&              a_ms_ks_modes,
-                                      hiptensorDataType_t                      typeB,
-                                      std::vector<std::size_t> const&          b_ns_ks_lengths,
-                                      std::vector<std::size_t> const&          b_ns_ks_strides,
-                                      std::vector<int32_t> const&              b_ns_ks_modes,
-                                      hiptensorDataType_t                      typeD,
-                                      std::vector<std::size_t> const&          d_ms_ns_lengths,
-                                      std::vector<std::size_t> const&          d_ms_ns_strides,
-                                      std::vector<int32_t> const&              d_ms_ns_modes,
-                                      hiptensorDataType_t                      typeE,
-                                      std::vector<std::size_t> const&          e_ms_ns_lengths,
-                                      std::vector<std::size_t> const&          e_ms_ns_strides,
-                                      std::vector<int32_t> const&              e_ms_ns_modes,
-                                      hiptensorComputeDescriptor_t             computeType,
-                                      const uint64_t                           workspaceSize);
+    hiptensorStatus_t bruteForceModel(ContractionSolution**              winner,
+                                      std::vector<ContractionSolution*>& candidates,
+                                      hiptensorDataType_t                typeA,
+                                      std::vector<std::size_t> const&    a_ms_ks_lengths,
+                                      std::vector<std::size_t> const&    a_ms_ks_strides,
+                                      std::vector<int32_t> const&        a_ms_ks_modes,
+                                      hiptensorDataType_t                typeB,
+                                      std::vector<std::size_t> const&    b_ns_ks_lengths,
+                                      std::vector<std::size_t> const&    b_ns_ks_strides,
+                                      std::vector<int32_t> const&        b_ns_ks_modes,
+                                      hiptensorDataType_t                typeD,
+                                      std::vector<std::size_t> const&    d_ms_ns_lengths,
+                                      std::vector<std::size_t> const&    d_ms_ns_strides,
+                                      std::vector<int32_t> const&        d_ms_ns_modes,
+                                      hiptensorDataType_t                typeE,
+                                      std::vector<std::size_t> const&    e_ms_ns_lengths,
+                                      std::vector<std::size_t> const&    e_ms_ns_strides,
+                                      std::vector<int32_t> const&        e_ms_ns_modes,
+                                      hiptensorComputeDescriptor_t       computeType,
+                                      const uint64_t                     workspaceSize);
 
     template <typename A,
               typename B,
@@ -64,7 +64,8 @@ namespace hiptensor
     struct ActorCriticSelection
     {
         static hiptensorStatus_t
-            selectWinner(ContractionSolution**                                   winner,
+            selectWinner(const hiptensorHandle_t                                 handle,
+                         ContractionSolution**                                   winner,
                          std::unordered_map<size_t, ContractionSolution*> const& candidates,
                          hiptensorDataType_t                                     typeA,
                          std::vector<std::size_t> const&                         a_ms_ks_lengths,
@@ -86,7 +87,8 @@ namespace hiptensor
     };
 
     hiptensorStatus_t
-        actorCriticModel(ContractionSolution**                                   winner,
+        actorCriticModel(const hiptensorHandle_t                                 handle,
+                         ContractionSolution**                                   winner,
                          std::unordered_map<size_t, ContractionSolution*> const& candidates,
                          hiptensorDataType_t                                     typeA,
                          std::vector<std::size_t> const&                         a_ms_ks_lengths,
@@ -108,4 +110,3 @@ namespace hiptensor
                          const uint64_t                                          workspaceSize);
 
 } // namespace hiptensor
-