+++ projects/rocwmma/samples/common.hpp
@@ -109,7 +109,9 @@ bool isGfx12()
     std::string deviceName(mProps.gcnArchName);
 
     return ((deviceName.find("gfx1200") != std::string::npos)
-            || (deviceName.find("gfx1201") != std::string::npos));
+            || (deviceName.find("gfx1201") != std::string::npos)
+            || (deviceName.find("gfx1250") != std::string::npos)
+            || (deviceName.find("gfx1251") != std::string::npos));
 }
 
 // HIP Host function to find if the device supports f64
@@ -125,12 +127,22 @@ bool isF64Supported()
 
     return ((deviceName.find("gfx90a") != std::string::npos)
             || (deviceName.find("gfx942") != std::string::npos)
-            || (deviceName.find("gfx950") != std::string::npos));
+            || (deviceName.find("gfx950") != std::string::npos)
+            || (deviceName.find("gfx1251") != std::string::npos));
 }
 
 bool isF32Supported()
 {
-    return isGfx9();
+    hipDevice_t     mHandle;
+    hipDeviceProp_t mProps;
+
+    CHECK_HIP_ERROR(hipGetDevice(&mHandle));
+    CHECK_HIP_ERROR(hipGetDeviceProperties(&mProps, mHandle));
+
+    std::string deviceName(mProps.gcnArchName);
+
+    return isGfx9() || (deviceName.find("gfx1250") != std::string::npos)
+           || (deviceName.find("gfx1251") != std::string::npos);
 }
 
 inline double calculateGFlops(uint32_t m, uint32_t n, uint32_t k)