+++ projects/hiptensor/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Full documentation for hipTensor is available at [rocm.docs.amd.com/projects/hiptensor](https://rocm.docs.amd.com/projects/hipTensor/en/latest/index.html).
 
+## (Unreleased) hipTensor 2.x
+
+### Added
+
+* Added support for new GPU target gfx1250.
+
 ## (Unreleased) hipTensor 2.3.0
 
 ### Added